+++ Sources/Models/DetectorType.swift
@@ -30,6 +30,7 @@ public enum DetectorType {
     case date
     case phoneNumber
     case url
+    case transitInformation
 
     // MARK: - Not supported yet
 
@@ -43,6 +44,7 @@ public enum DetectorType {
         case .date: return .date
         case .phoneNumber: return .phoneNumber
         case .url: return .link
+        case .transitInformation: return .transitInformation
         }
     }